+++ packages/universal-header/CHANGELOG.md
@@ -3,6 +3,74 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.9-rc.7](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.2.9-rc.6...@twreporter/universal-header@2.2.9-rc.7) (2022-06-06)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
+## [2.2.9-rc.6](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.2.9-rc.5...@twreporter/universal-header@2.2.9-rc.6) (2022-05-30)
+
+
+### Bug Fixes
+
+* clean members to null ([116fd6a](https://github.com/twreporter/twreporter-npm-packages/commit/116fd6a51a00daec61a6d80c4ddb4121e33b5ffd))
+* fix header part ([b5afe9b](https://github.com/twreporter/twreporter-npm-packages/commit/b5afe9bec836c5806377edf09592daae23cfd03a))
+
+
+
+
+
+## [2.2.9-rc.5](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.2.9-rc.4...@twreporter/universal-header@2.2.9-rc.5) (2022-05-27)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
+## [2.2.9-rc.4](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.2.9-rc.3...@twreporter/universal-header@2.2.9-rc.4) (2022-05-27)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
+## [2.2.9-rc.3](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.2.9-rc.2...@twreporter/universal-header@2.2.9-rc.3) (2022-05-25)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
+## [2.2.9-rc.2](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.2.9-rc.1...@twreporter/universal-header@2.2.9-rc.2) (2022-05-25)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
+## [2.2.9-rc.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.2.9-rc.0...@twreporter/universal-header@2.2.9-rc.1) (2022-05-25)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
+## [2.2.9-rc.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.2.8...@twreporter/universal-header@2.2.9-rc.0) (2022-05-12)
+
+**Note:** Version bump only for package @twreporter/universal-header
+
+
+
+
+
 ## [2.2.8](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/universal-header@2.2.8-rc.1...@twreporter/universal-header@2.2.8) (2022-05-09)
 
 **Note:** Version bump only for package @twreporter/universal-header